restructured and refactored classes and labs
