+++ mlflow/extras/lab_utils.py
@@ -1,5 +1,3 @@
-import statistics
-
 import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
