+++ mlflow/labs/02_banknote_classification_lab.py
@@ -30,106 +30,8 @@ https://stackabuse.com/random-forest-algorithm-with-python-and-scikit-learn/
 """
 
 import mlflow.sklearn
-from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import StandardScaler
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.metrics import classification_report, confusion_matrix, accuracy_score, precision_score
-from lab_utils import Utils
-
-class RFCModel():
-
-    def __init__(self, params={}):
-        """
-        Constructor for RandomForestClassifier
-        :param params: parameters for the constructor such as no of estimators, depth of the tree, random_state etc
-        """
-        self.rf = RandomForestClassifier(**params)
-        self._params = params
-
-    @classmethod
-    def new_instance(cls, params={}):
-        return cls(params)
-
-    @property
-    def model(self):
-        """
-        Fetch the model
-        :return: return the model
-        """
-        return self.rf
-
-    @property
-    def params(self):
-        """
-        getter for model parameters
-        """
-        return self._params
-
-    def mlflow_run(self, df, r_name="Lab-2:RF Bank Note Classification Experiment"):
-        """
-        This method trains, computes metrics, and logs all metrics, parameters,
-        and artifacts for the current run
-        :param df: pandas dataFrame
-        :param r_name: Name of the experiment as logged by MLflow
-        :return: MLflow Tuple (ExperimentID, runID)
-        """
-
-        with mlflow.start_run(run_name=r_name) as run:
-            # get all attributes
-            X = df.iloc[:, 0:4].values
-            # get all the last columns, which is what we want to predict, our values
-            y = df.iloc[:, 4].values
-
-            # create train and test data
-            X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)
-
-            # Feature Scaling
-            sc = StandardScaler()
-            X_train = sc.fit_transform(X_train)
-            X_test = sc.transform(X_test)
-
-            # train and predict
-            self.rf.fit(X_train, y_train)
-            y_pred = self.rf.predict(X_test)
-
-            # Log model and params using the MLflow sklearn APIs
-            mlflow.sklearn.log_model(self.model, "random-forest-class-model")
-            mlflow.log_params(self.params)
-
-            # compute evaluation metrics
-            acc = accuracy_score(y_test, y_pred)
-            precision = precision_score(y_test, y_pred)
-
-            # log metrics
-            mlflow.log_metric("accuracy_score", acc)
-            mlflow.log_metric("precision", precision)
-
-            # get current run and experiment id
-            runID = run.info.run_uuid
-            experimentID = run.info.experiment_id
-
-            # create confusion matrix images
-            (plt, fig, ax) = Utils.plot_confusion_matrix(y_test,y_pred,y, title="Bank Note Classification Confusion Matrix")
-
-            # create temporary artifact file name and log artifact
-            temp_file_name = Utils.get_temporary_directory_path("confusion_matrix-", ".png")
-            temp_name = temp_file_name.name
-            try:
-                fig.savefig(temp_name)
-                mlflow.log_artifact(temp_name, "confusion_matrix_plots")
-            finally:
-                temp_file_name.close()  # Delete the temp file
-
-            # print some data
-            print("-" * 100)
-            print("Inside MLflow Run with run_id {} and experiment_id {}".format(runID, experimentID))
-            print("Estimators trees:", self.params["n_estimators"])
-            print(confusion_matrix(y_test,y_pred))
-            print(classification_report(y_test,y_pred))
-            print("Accuracy Score:", acc)
-            print("Precision     :", precision)
-
-            return (experimentID, runID)
+from lab_cls.rfc_model import RFCModel
+from lab_cls.lab_utils import Utils
 
 #
 # Lab/Homework for Some Experimental runs
@@ -143,6 +45,7 @@ class RFCModel():
 
 if __name__ == '__main__':
     # load and print dataset
+    mlflow.set_tracking_uri("sqlite:///mlruns.db")
     dataset = Utils.load_data("data/bill_authentication.csv")
     Utils.print_pandas_dataset(dataset)
     # iterate over several runs with different parameters
