+++ mlflow/solutions/03_lab.py
@@ -26,77 +26,10 @@ https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestR
 https://towardsdatascience.com/understanding-random-forest-58381e0602d2
 https://github.com/MangoTheCat/Modelling-Airbnb-Prices
 """
-
-import mlflow.sklearn
-import numpy as np
-from sklearn.model_selection import train_test_split
-from sklearn.ensemble import RandomForestRegressor
-from sklearn import metrics
-from lab_utils import Utils
-
-class RFRBaseModel():
-
-    def __init__(self, params={}):
-        """
-        Construtor for the RandomForestRegressor
-        :param params: dictionary to RandomForestRegressor
-        """
-        self.params = params
-        self.rf = RandomForestRegressor(**params)
-
-    @classmethod
-    def new_instance(cls, params={}):
-        return cls(params)
-
-    def model(self):
-        """
-        Getter for the model
-        :return: return the model
-        """
-        return self.rf
-
-    def mlflow_run(self, df, r_name="Lab-3: Baseline RF Model"):
-        """
-        This method trains, computes metrics, and logs all metrics, parameters,
-        and artifacts for the current run
-        :param df: pandas dataFrame
-        :param r_name: Name of the experiment as logged by MLflow
-        :return: Tuple of MLflow experimentID, runID
-        """
-        with mlflow.start_run(run_name=r_name) as run:
-            X_train, X_test, y_train, y_test = train_test_split(df.drop(["price"], axis=1), df[["price"]].values.ravel(), random_state=42)
-            self.rf.fit(X_train, y_train)
-            predictions = self.rf.predict(X_test)
-
-            # Log model and parameters
-            mlflow.sklearn.log_model(self.rf, "random-forest-model")
-            mlflow.log_params(self.params)
-
-            # Create metrics
-            mae = metrics.mean_absolute_error(y_test, predictions)
-            mse = metrics.mean_squared_error(y_test, predictions)
-            rsme = np.sqrt(mse)
-            r2 = metrics.r2_score(y_test, predictions)
-
-            # Log metrics
-            mlflow.log_metric("mae", mae)
-            mlflow.log_metric("mse", mse)
-            mlflow.log_metric("rsme", rsme)
-            mlflow.log_metric("r2", r2)
-
-            runID = run.info.run_uuid
-            experimentID = run.info.experiment_id
-
-            # print some data
-            print("-" * 100)
-            print("Inside MLflow {} Run with run_id {} and experiment_id {}".format(r_name, runID, experimentID))
-            print("Estimator trees        :", self.params["n_estimators"])
-            print('Mean Absolute Error    :', mae)
-            print('Mean Squared Error     :', mse)
-            print('Root Mean Squared Error:', rsme)
-            print('R2                     :', r2)
-
-            return (experimentID, runID)
+import mlflow
+from sol_cls.rfr_base_exp_model import RFRBaseModel
+from sol_cls.lab_utils import Utils
+#
 #
 # TODO in Lab/Homework for Some Experimental runs
 #
@@ -105,6 +38,8 @@ class RFRBaseModel():
     # 3. Check in MLflow UI for parameters, metrics, and artifacts
 
 if __name__ == '__main__':
+    # Use sqlite:///mlruns.db as the local store for tracking and registry
+    mlflow.set_tracking_uri("sqlite:///mlruns.db")
     # load and print dataset
     dataset = Utils.load_data("data/airbnb-cleaned-mlflow.csv")
     Utils.print_pandas_dataset(dataset)
